Fix tests and screenshots
