+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/ConversationScreen.kt
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.VerticalDivider
-import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -16,8 +15,8 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import de.tum.informatics.www1.artemis.native_app.core.ui.getWindowSizeClass
 import de.tum.informatics.www1.artemis.native_app.core.ui.navigation.DefaultTransition
+import de.tum.informatics.www1.artemis.native_app.core.ui.useTabletLayout
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.StandalonePostId
 import org.koin.androidx.compose.koinViewModel
 import org.koin.core.parameter.parametersOf
@@ -53,7 +52,7 @@ fun ConversationScreen(
         mutableStateOf(threadPostId != null)
     }
 
-    WindowSizeAwareTwoColumnLayout(
+    LayoutAwareTwoColumnLayout(
         modifier = modifier,
         optionalColumn = conversationsOverview,
     ) { innerModifier ->
@@ -94,17 +93,15 @@ fun ConversationScreen(
 
 
 @Composable
-fun WindowSizeAwareTwoColumnLayout(
+fun LayoutAwareTwoColumnLayout(
     modifier: Modifier = Modifier,
     optionalColumnWeight: Float = 1f,
     priorityColumnWeight: Float = 2f,
     optionalColumn: @Composable (Modifier) -> Unit,
     priorityColumn: @Composable (Modifier) -> Unit
 ) {
-    val widthSizeClass = getWindowSizeClass().widthSizeClass
-
     when {
-        widthSizeClass <= WindowWidthSizeClass.Compact -> {
+        !useTabletLayout() -> {
             Box(modifier = modifier) {
                 priorityColumn(Modifier.fillMaxSize())
             }
