--- feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/post/PostActionFlags.kt
@@ -1,7 +1,0 @@
-package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.post
-
-data class PostActionFlags(
-    val isAbleToPin: Boolean,
-    val isAtLeastTutorInCourse: Boolean,
-    val hasModerationRights: Boolean
-)
